+++ src/logger.h
@@ -1,10 +1,11 @@
 #ifndef LOGGER_H
 #define LOGGER_H
 
+#include <stddef.h>
+
 extern void logger_errors_write(const char* fmt, ...);
 extern void logger_requests_write(const char* fmt, ...);
 extern void logger_changes_write(const char* fmt, ...);
-
-extern int logger_changes_sumtime(char* buffer, int buffer_size, const char* pattern);
+extern size_t logger_changes_sumtime(char* buffer, size_t buffer_size, const char* pattern);
 
 #endif // LOGGER_H
